+++ robotcode/utils/debugpy.py
@@ -1,3 +1,5 @@
+from typing import Sequence, Tuple, Union
+
 from .logging import LoggingDescriptor
 from .net import find_free_port
 
@@ -24,24 +26,34 @@ def wait_for_debugpy_connected() -> bool:
     return False
 
 
-def enable_debugpy(port: int) -> bool:
+def enable_debugpy(port: int, addresses: Union[Sequence[str], str, None] = None) -> bool:
     if is_debugpy_installed():
         import debugpy
 
-        debugpy.listen(port)
+        if addresses is None:
+            addresses = ["127.0.0.1"]
+
+        if not isinstance(addresses, Sequence):
+            addresses = [addresses]
+
+        for address in addresses:
+            debugpy.listen((address, port))
 
         return True
     return False
 
 
-def start_debugpy(port: int, wait_for_client: bool) -> bool:
+def start_debugpy(end_point: Union[Tuple[str, int], int], wait_for_client: bool) -> bool:
     if is_debugpy_installed():
         import debugpy
 
-        real_port = find_free_port(port)
-        if real_port != port:
+        if isinstance(end_point, int):
+            end_point = ("127.0.0.1", end_point)
+
+        real_port = find_free_port(end_point[1])
+        if real_port != end_point[1]:
             _logger.warning(f"start debugpy session on port {real_port}")
-        debugpy.listen(real_port)
+        debugpy.listen((end_point[0], real_port))
 
         if wait_for_client:
             _logger.info("wait for debugpy client")